requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
psycopg2-binary==2.9.5
celery==5.2.7
redis==4.5.1
djangorestframework_simplejwt==4.8.0
drf-spectacular==0.26.0
Pillow==9.4.0
django_filter==22.1